pyproject.toml
[project]
name = "lmeval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "openai>=1.69.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "tqdm>=4.67.1",
]
